requirements.txt
# Django
django==4.2.1
django_environ==0.10.0
mysqlclient==2.0.3
djangorestframework==3.14.0
django-cors-headers==3.10.1
pytz==2021.3
django-sslserver==0.22
pyOpenSSL==22.1.0
django-simpleui==2023.8.28

# Storage
redis==4.5.4
django-redis==5.2.0

# Celery
celery==5.2.2

# Server
uwsgi==2.0.21

# Datetime
arrow==1.2.3

# OT
protobuf==3.19.5
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-django==0.40b0
opentelemetry-instrumentation-dbapi==0.40b0
opentelemetry-instrumentation-redis==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-celery==0.40b0
opentelemetry-instrumentation-logging==0.40b0

# Log
python_json_logger==2.0.4

# RSA
pycryptodome==3.15.0

# Request
requests==2.27.1

# Profile
pyinstrument==4.4.0
